done user update

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,9 +19,9 @@ class CreateUsersTable extends Migration
             $table->string('username')->unique();
             $table->string('email')->unique();
             $table->string('password');
-            $table->integer('npak')->nullable();
+            $table->bigInteger('npak')->nullable();
             $table->enum('level', ['Ketua SPI', 'Auditor', 'Auditee', 'Direktur']);
-            $table->integer('nohp')->nullable();
+            $table->bigInteger('nohp')->nullable();
             $table->string('foto')->nullable();
             $table->string('ttd')->nullable();
             $table->enum('status', ['Aktif', 'Tidak Aktif', 'Medaftar']);
